server/requirements.txt
numpy
pandas
matplotlib
feedparser
requests
beautifulsoup4
pydantic
python-dotenv
fastapi
uvicorn
openai
scikit-learn
lxml